Removing util.py. Adding datasets module with only MNIST and 'generate_spike_train()' for now.
