+++ examples/eth.py
@@ -16,10 +16,10 @@ from mpl_toolkits.axes_grid1 import make_axes_locatable
 sys.path.append(os.path.abspath(os.path.join('..', 'spiketorch')))
 sys.path.append(os.path.abspath(os.path.join('..', 'spiketorch', 'network')))
 
-from util import *
-from network import Network
+from network import *
 from monitors import Monitor
 from synapses import Synapses, STDPSynapses
+from datasets import get_MNIST, generate_spike_train
 from groups import InputGroup, LIFGroup, AdaptiveLIFGroup
 
 model_name = 'eth'
@@ -182,17 +182,13 @@ network.add_synapses(Synapses(network.groups['Ai'], network.groups['Ae'], w=-c_i
 # network.add_monitor(Monitor(obj=network.groups['Ae'], state_vars=['v', 'theta']), name=('Ae', ('v', 'theta')))
 # network.add_monitor(Monitor(obj=network.groups['Ai'], state_vars=['v']), name=('Ai', 'v'))
 
-# Get training, test data from disk.
+# Get training or test data from disk.
 if mode == 'train':
-	data = get_labeled_data('train', train=True)
+	data = get_MNIST(train=True)
 elif mode == 'test':
-	data = get_labeled_data('test', train=False)
+	data = get_MNIST(train=False)
 
-# Convert data into torch Tensors.
-if mode == 'train':
-	X, y = data['X'], data['y']
-elif mode == 'test':
-	X, y = data['X'], data['y']
+X, y = data['X'], data['y']
 
 # Count spikes from each neuron on each example (between update intervals).
 spike_monitor = np.zeros([update_interval, n_neurons])
